requirements.txt
alembic==1.7.7
anyio==3.6.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pydantic==1.9.0
python-decouple==3.6
hubspot-api-client==5.0.0
requests==2.27.1
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
stripe==3.0.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
